requirements.txt
djangorestframework
absl-py==0.8.1
affine==2.3.0
alembic==1.3.3
amqp==2.5.2
APScheduler==3.6.3
arrow==0.15.5
astor==0.8.1
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.1.0
bleach==3.1.0
blinker==1.4
boto==2.49.0
cachetools==4.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
checksumdir==1.1.7
circus==0.16.1
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.2.2
colorlog==4.1.0
cryptography==2.8
cycler==0.10.0
deap==1.3.0
decorator==4.4.1
descartes==1.1.0
diff-match-patch==20181111
distro==1.4.0
dnspython==1.16.0
docker==4.1.0
earthengine-api==0.1.210
earthpy==0.8.0
elasticsearch==7.5.1
elasticsearch-dsl==7.0.0
envparse==0.2.0
eventlet==0.25.1
fabric==2.5.0
factory-boy==2.12.0
Faker==4.0.0
Fiona==1.8.13
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Caching==1.8.0
Flask-CeleryExt==0.3.2
Flask-Cors==3.0.8
Flask-Elasticsearch==0.2.5
Flask-JWT-Extended==3.24.1
Flask-Limiter==1.1.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-REST-JSONAPI==0.12.6
Flask-Script==2.0.6
Flask-Scrypt==0.1.3.6
Flask-SQLAlchemy==2.4.1
forex-python==1.5
future==0.18.2
gast==0.2.2
GeoAlchemy2==0.6.3
geoip2==3.0.0
geojson==2.5.0
geojsonio==0.0.3
geopandas==0.6.2
github3.py==1.3.0
greenlet==0.4.15
grpcio==1.25.0
gunicorn==20.0.4
gym==0.14.0
h5py==2.10.0
healthcheck==1.3.3
html5lib==1.0.1
httplib2==0.16.0
humanize==0.5.1
icalendar==4.0.4
idna==2.8
imageio==2.6.1
importlib-metadata==1.4.0
invoke==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
jsonpickle==1.2
jwcrypto==0.6.0
kiwisolver==1.1.0
kombu==4.6.7
lgblkb-tools==0.3.10.4
limits==1.5
Mako==1.1.1
Markdown==3.1.1
MarkupSafe==1.1.1
marrow.mailer==4.0.3
marrow.util==1.2.3
marshmallow==2.15.6
marshmallow-jsonapi==0.22.0
matplotlib==3.1.2
maxminddb==1.5.2
monotonic==1.5
more-itertools==7.2.0
munch==2.5.0
networkx==2.3
numexpr==2.7.0
numpy==1.17.4
oauth2==1.9.0.post1
omise==0.8.1
opt-einsum==3.1.0
ortools==7.3.7083
pandas==0.25.3
paramiko==2.7.1
paypalrestsdk==1.13.1
Pillow==6.1.0
protobuf==3.11.2
psutil==5.6.7
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
pycparser==2.19
pycryptodome==3.9.4
pyglet==1.3.2
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyPDF2==1.26.0
pyproj==2.4.2.post1
python-box==4.0.4
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
python-geoip==1.2
python-geoip-geolite2==2015.303
python-http-client==3.2.4
python-log-indenter==0.9
python-magic==0.4.15
python-pentabarf-xml==0.20
python-telegram-bot==12.3.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
qrcode==6.1
rasterio==1.1.2
redis==3.3.11
reportlab==3.5.34
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.0
scrypt==0.8.13
sendgrid==6.1.1
sentry-sdk==0.14.1
Shapely==1.6.4.post2
simplejson==3.17.0
six==1.13.0
sklearn==0.0
snuggs==1.4.7
sortedcontainers==2.1.0
SQLAlchemy==1.3.12
SQLAlchemy-Continuum==1.3.9
SQLAlchemy-Utils==0.36.1
stripe==2.42.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.0
tornado==4.5.3
typing==3.7.4.1
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
visvalingamwyatt==0.1.2
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xxhash==1.4.3
zipp==2.1.0
